Fix redlock deprecation warnings
**ISSUE**
After the `Redis.current` fix, we are now seeing significant numbers
of deprecations warnings like:
```
Passing 'call' command to redis as is; blind passthrough has been
  deprecated and will be removed in redis-namespace 2.0
  (at /Users/mark/.rbenv/versions/2.7.4/lib/ruby/gems/2.7.0/gems/
   redlock-2.0.6/lib/redlock/client.rb:214:in `block (2 levels) in unlock')
```

**DIAGNOSIS**
We were passing a RedisNamespace object to Redlock instead of a
plain Redis connection specifier. RedisNamespace was complaining
any time an admin (not-namespace-able) command was received.

**RESOLUTION**
Pass a Redis URI to RedLock instead of a RedisNamespace object.