Speed up AttachFilesToWorkJob test
**ISSUE**
The test was creating a new ETD and file, and then running the
job for each example, and then only reading the results. I.E. we
were running expensive setup multiple times with no variation and
only reading (not modifying or acting on) the output.

**RESOLUTION**
Doing the setup and running the job only once reduces the test
time by over 30% (avg 27.9s down to avg 9.8s).